C3M1: Tweaks to Prisoner Camp Wanderers
- Added new A_WanderGoal function to the Base class.  If the actor has a
goal set, it will walk to that goal.  Once the goal is reached, or if it
has no goal, it will wander normally.
- Also tweaked the Camp_PrisonerBald actor to no longer use A_Look2 -
A_Look2 doesn't execute specials, so using Thing_SetGoal wouldn't work.
- Fixed all Thing_SetGoal calls on the prisoner actors also - first two
arguments were reversed

NOTE: The map has some kind of weird issue where there is a "dead zone"
just outside of the train car...  The prisoners there get stuck and
don't move, and any prisoners that enter the area get stuck.  I tried to
find what was causing it, but no luck fixing it...  Deleting large
chunks of the area of the map around the crematorium area fixed it for
me, but I couldn't narrow the issue down to a single sector, so it's
still broken here.