Initial ZScript Implementation of Actor Rolling/Dodge
- Added the A_DoSideRoll function to the Base class.  See WGuard actor
for use. First parameter (default 10) controls the speed to dodge at.
- Checks for blocking items or walls to the left or the right of the
actor, identifies the side with the most room, and rolls to that side.
Should handle avoiding dropoffs as well, but needs more testing.
- Also added a variable to save the initial sprite of the actor, so that
resetting to default can be done via ZScript as needed.  Added handling
for this to the BecomeIdle funcitno fo stealth actors as well, so they
would revert to their normal sprite when they returned to idle, even if
mid-roll