C1M2
Mutant Labs : new additions to improve gameplay.

Green door is now further in the level, save player wasting time.

Bigger door leading to the semi-circle room, so the mutants can now be activated behind the glass and be in full shot.

New "exit" in underwater area to save player time.