Dodging Enemy Invulnerability Fix
- Trying to fix this again...  Forcing Invulnerability to be disabled in
the "See.Dodge" state, so that if the Dodge.End state gets skipped
somehow, the enemy will end up vulnerable in the next tic of wandering
anyway
- Also made the size change a modifier of Default height instead of just
multiplying/dividing directly, to be sure to avoid any recursive height
reduction if this somehow gets called twice.

I've massacred the enemies in the test map over and over again, and
never got any ghosts with this fix - please test and confirm.