Added initial implementation of SkyViewPointStatic actor
- When used as a Skybox Viewpoint, keeps the skybox scene static
relative to the player.  If there are multiple players, acts just like a
normal SkyViewpoint actor.
- Functions identically to the SkyViewPoint actor, with the addition of
optional Arg 1 to set the relative scale of the skybox scene (The larger
this number is, the farther away the skybox contents will appear to be).
- See demonstration of use in WOLFNEW2 map - I just replaced the
SkyViewpoint actor with this one, and set argument 1 to 60