SkyViewPointStatic Tweak
- Added NOCLIP flag to the SkyViewPointStatic actor and changed the
height/radius to zero so that it won't get stuck on other actors or
level gemoetry.

NOTE: Be careful when placing structures/hills/etc. in a skybox that
uses a static viewpoint...  The viewpoint actor needs to be able to move
around in the skybox in order to keep the view positioned properly for
the player.  Like all actors, though, the viewpoint cannot move through
floors, so if it encounters level geometry with a height change, the
z-height of the viewpoint won't be adjusted properly, and the skybox
view height will be incorrect for the player's position.