SkyViewPointStatic Re-work
- Added "anchor actor" handling to the SkyViewPointStatic actor.  This
lets you anchor a skybox to an actor (normally a static mapspot) with a
specific TID instead of to the player's spawn point (leave arg2 as 0 for
original behavior).  See C3M5_C airship bridge for an example (anchor is
on the captain's chair).  Previously, moving the player spawn to another
location, then using NOCLIP to get to the bridge would result in the
skybox being completely misaligned because it was centered on the
player's spawn location.
- Changed positioning code to use coordinate offsets directly instead of
velocity.  This means that the viewpoint actor will now always move in
concert with the player's x/y movement, regardless of if the viewpoint
hits an impassible line or not, so will stay properly offset/aligned.
- Also added SkyViewPointAnchor actor to act as a dedicated anchor point
actor class.