C3M1: Did some work on the underground area
Added a huge explosion to the unfinished structures, just start the map,
enter the right door and use the switch there to test it. I timed sound
and visuals to make it feel as the explosion is farther away. What would
be cool @AFADoomer (maybe you can come up with something) is some kind
of a blurry sight (shader?) for the player just to make him feel as the
shockwave of the explosion hit him hard. The blurry sight can fade to
normal within 2-3 seconds - we could even add this to the UFO explosion
in C2M5_B or a weaker version in case you get some splash damage from
grenades or rockets? ...just brainstorming...