Added Inventory-controlled blur effect
- Added a simple blur shader that can be controlled by giving the player
the "BlurShaderControl" inventory item via script
- Can be tested with: "give blurshadercontrol 175" in console (sets blur
for 5 seconds)
- Is handled using the same base code as the "old video" shader and
shake shader items, with the addition that the amount of the item that
you give sets how long the effect lasts - if you give 35, the effect
will last 35 tics, 70 -> ~2 seconds, 350 -> ~10 seconds, etc.  The
effect will stack if you give more when there's already a blur in
action.
- There's also a ~0.5 second fade in to the blur effect, then a slow
fade back to normal over time once the blur is in place