More healing/medic/resurrection code changes
Added flags to control how medic/Totengraber healing works....  Setting
Nazi.Healer to 1 (defined as HLR_ALLIES) is normal medic behavior.
Nazi.Healer 2 (HLR_ZOMBIES) changes the behavior to only search for dead
enemies that have a zombie counterpart, and to replace the dead actor
with the zombie variant on resurrection.
- To demo: Spawn Totengraber_Wounded and WGuard.  Kill WGuard, and he'll
be resurrected by Totengraber as a ZombieBrain.  See base.txt, line 1064
for the list of "zombie-able" classes that can be respawned, and line
1340 for which regular enemies become what zombie.
- Also fixed a frame definition error in the Camp_PrisonerBald Pain
state that caused the shirtless variant to use the wrong death frames