Spotlight and AlertLight tweaks
- Updated spotlight visibility handling to accurately calculate the
player's distance from the center of the spotlight beam, taking into
account player radius and height when calculating relative pitch and
angle.  This is actually more accurate than the standard light handling,
which uses the actor's position directly and approximates light level
from there - but the precision is necessary here in order to be more
realistic with the effect.
- Tweaked normal alert lights to still function when the level is
frozen, and removed some DECORATE-isms in the code.