Enemy Light Avoidance Handling Test
- Added function to the Base class that calculates the visibility of the
enemy actor, based on the dynamic lights that are in range of the actor
(doesn't account for spotlights)
- Right now, it's only used and integrated into the WaveSpawners...  You
can see the effect in action in C3M0_A - if you trigger the first wave,
then stand in the middle near the campfire, the zombies will not walk
into the red/orange lighted area.

This can be tweaked, still.  Or removed, if it causes unwanted
effects...  But it's only active on the secret level right now, anyway,
and only for WaveSpawned enemies.