C3M0_A: Work on the two areas.
I started to give each area a certain mission.

The campfire area's mission is to collect the skull key, enter the moon
shelter (door is missing for key) and get the shovel to be equipped with
something more useful against the melee zombies.

After that (and another night), the player is suppoed to reach the
mansion, enter the mansion, get the repair kit, repar all three power
stations, reactivate the light, enter the cellar and find the lantern.

As of now, the whole gameplay doesn't seem to work well. I can't tell if
it is too complicated in general (to use, to play, to set up) or if the
day night cycle isn't working as it is supposed.

I think it should be easier for the player to understand how the main
mechanic works (e.g. missions as fast as possible during daytime, but
also at nighttime just also trying to defend against Zombies.

I have no idea on how to improve the gameplay mechanics currently but
maybe @AFADoomer you have some good ideas... maybe try what I have
created so far.