Modified disguise-friendly weapon handling
- Added a new Stealth property to the NaziWeapon class.  Any weapon that
inherits from NaziWeapon and has NaziWeapon.Stealth set to 1 can be
carried while the player is disguised, and won't break the disguise
unless the weapon is fired.
- Used this property to significantly simplify the NullWeapon code and
to make the Shovel carryable while disguised (which makes sense, since
there are prisoners with shovels already).  As long as you aren't seen
attaacking with it, the disguise will hold.
- Also updated status bar handling accordingly to remove hard-coded
NullWeapon reference and adjust display of visibility bar appropriately
if the 'stealth' weapon is fired while you are disguised