C3M0_A: Code tweaks
- Restored behavior of new waves starting immediately when triggered at
night (so if you enter the house area at night, the enemy spawns will
trigger as soon as you cross the line)
- Tweaked light avoidance behavior so that enemies will avoid bright
lights *unless* they are within 64 units of their target, in which case
they will charge in and attack regardless of light level.  TODO: I still
need to add handling for spotlights to the light level calculation.
- Did wome other minor cleanup