"Soul Gate" experiment for C3M0_A
- Added actor that performs inventory checks and acts as a
proximity-activated line blocker/unblocker.
- The GhostGate actor can be placed on a line, and turns that line into
an impassible swirling portal that spews ghosts that cause damage to
players in range who do not carry the necessary amount of the 'Soul'
inventory item to pass.  The effects fade away to nothing when a player
with enough Soul items approaches.
- The amount of 'Soul' needed can be specified in the first argument of
the map object (default is 100).

See the CreepySpawn room in the test map, or the cemetery entrance in
C3M0_A for an example...  The cemetery gate is set to require 10 Souls
in inventory to pass ('give soul 10' in console).