Tweaks and cleanup
- Split the Soul inventory items out into their own file
- Added a "StackableInventory" base class so that you can have
Ammo-style stacking of inventory items that share inheritance.  This is
almost not worth the extra file, because you can just inherit from Ammo
and add a couple of lines in your Defaults block and do the same thing
(See CoinItem), but...  easily done, and slightly cleaner.
- Added handling to GhostGate to ensure that all spirits are removed
when you "open" the gate, and tweaked a couple of other minor things.