Soul Gate additions, C3M0_A tweaks
- Added some error messages that show if you set up the gate improperly
- Added handling so that gates with the same TID are 'linked' together;
if one is deactivated, the others will be deactivated as well.
- Added a 'helper' actor that's basically a soul gate with no effects,
so that you can block multiple lines while still only having one main
SoulGate actor with all of the effects.  Note that they should be set up
with the same soul amounts as the main gate, otherwise they'll be
easier/harder to pass!

C3M0_A:
- Added the user_lightburns UDMF property to all of the outside
WaveSpawners, and updated ZScript to no longer hard-code the effect on
this level.
- Also removed the 801 TID argument from the underground wavespawners so
that just in case we bring back that piece of the day/night cycle
script, they won't be affected (the TID currently does nothing).
- The message ACS code was all reverted again in the last commit.  Maybe
I'll fix it again later.