Hold-to-activate update
- Added handling so that custom crosshairs (user_crosshair property) can
be set for 3D midtex lines and will only display while the crosshair is
over the midtex texture.

NOTE:  You *must* set the line to Block Hitscans for this to work!  If
it's set to Block Everything, the custom crosshair will be draw in open
air above/below the midtex, and if you don't set the line to either of
those two blocking types, the trace won't be able to see the line, and
nothing will draw (and hold-to-use won't work).