Soul Gate pairing re-work, additions
- Soul Gates (and all of their components) now respect in-editor scaling
- Soul Gate helpers now automatically get their key amount value from a
Soul Gate that has the same TID.
- Soul Gates with the same TID now synchronize their alpha values and
activation status, so paired gates will show the same activation
effects, even if the player is only near one of them (so the
swirl/light/spirits don't just disappear suddenly from the second gate
when the first gate is activated).
- Also added some quickie icons for the Soul Gate and helper, so they
are less obtrusive in the editor.