Light avoidance/death code cleanup and consolidation into Base class
- Moved light avoidance code into the main Base class.
- This adds two properties to Base-class descendants:
-- Base.LightThreshold, sets the light level at which the enemy will
turn around and run away (afraid of light)
-- Base.LightKills, which adds to the behavior above by causing any
enemy that is under a sky texture that breaks their defined light
threshold to die (via fire, if they have the state defined for it) - as
with the zombie deaths in C3M0_A.  Also causes smoke to spawn all over
the enemy when it's on the edge of bursting into flame.
- Both of these flags are set on spawned enemies by the WaveSpawners if
user_lightburns is set, but they could be placed on all of the zombie
enemies by default.  Would need to test for gameplay issues.
- Also made a couple of minor fixes to the spotlight code