C3M0_A: Hold-to-use script tweaks
- Added a little bit of logic so that "repair" sounds are only played at
specified intervals
- This uses the internal "repairstate[]" values that are keeping track
of the repair status as a counter of sorts, since the value is
incremented every tick while the player holds use...  A line with id of
117 has its hold time stored in "repairstate[117]", so checking that
value lets us delay re-playing sounds until the previous iteration is
done playing.
- Updated all of the previous uses of this code to use the same logic to
limit the number of sounds spammed.

@Tormentor667 - One recommendation: the last parameter of the 'Activate'
ACS script call is the number of seconds that the line must be held
down....  Right now they're all set to 5 seconds; it might be better to
change them up a little bit, so not every activation point is exactly
the same.

Also, do you want me to do a 3D model of the crank arm so that we can
have it stick out of the wall and rotate as we press use?