Supply Chest 3D Model/ZScriptification
- Replaces the SupplyChest actor and associated ACS with a
BoASupplyChest ZScript actor that is an animated 3D model.
- Only functional difference from the sprite version is: Instead of
directly giving the chest contents to the player, those contents are
actually spawned "inside" the opened chest for the player to pick up.
- This also lets map authors optionally specify a class name via the
string value of the first argument that will be spawned when the chest
is opened instead of one of the presets.

This started out as an experiment, and I got carried away...  I've left
all of the original Supply Chest code and assets in place for now, in
case we decide to revert (I just changed MAPINFO to point to the new
class).

This change will probably require adjusting most of the supply chests in
the game, since they were often placed with no regard to angle due to
being one-directional sprites at the time.