C3M0_A: Hold-to-use update/expansion, re-included message strings
- Added separate rendering call for icon crosshairs that is separate
from the actual crosshair, so that I can control color, opacity, etc.
separately from the player's real crosshair
- When configuring a linedef to have custom crosshair, the property type
can now be set to Text or Thing Class instead of Integer in order to
specify the inventory item that is needed to activate the line.  Note
that the old integer method will still work!  You just don't get any of
the added functionality or checks described below.
- Setting a string value on the line's property allows for changing the
crosshair effect based on if you have the named item in your inventory
or not, and for the inventory item's icon to be used as the crosshair
image, without adding additional XHAIR* files.

I've updated all of the hold-to-use lines in C3M0_A to reflect this
change.  When you hover over them intially, the icon crosshair will show
in orange and pulse, indicating that you don't have the needed item.
When you do have the item, the icon will be steady and green.  Once you
hit 'use', the status meter will show up in white/grey.

I also re-added the message strings to the ACS script again.

In order to allow us to use higher resolution crosshair icons, I also
re-worked the status bar's inventory bar drawing to force the size of
inventory icons to 32x32 or smaller.  I copied the old XHAIRB95 image
over the old Crank icon as a proof of concept - it's 44x52, but gets
scaled to 27x32 (or so) automatically.