C3M0_A: Added "text hints" system - far from being finished, need some help from AFADoomer
@AFADoomer I implemented a rudimentary text system which displays a
paper background (through hudmessage) when using an actor (flatsprite
paper actor). You need to hold "use" to read on.

Currently the text is static but it should be somehow changed to make
the script dynamic, so by simply passing a number or a string through
the script execute arguments, you can choose which text is displayed
(maybe by a string that corresponds to a LANGUAGE entry).

To test it, simply start the map and use the paper actor to your right.
I am pretty sure you can come up with something that works like a charm.

For what? Well, I want to place some text notes through various maps
from C1 to C3 that tell the story a bit more detailed for those who are
interested. Next to that - I don't want BJ to talk. We did a good job in
avoiding a talking "Duke Nukem like" protagonist/player in C1 and C2 and
I really don't to change this now in C3, it simply feels wrong. Giving
clues through texts seems to be a better choice, also feels very
"Resident Evil"'ish in C3M0_A :)