Tex Hints Extensions
- Exported messages to LANGUAGE - you pass in a message number (e.g.,
5), and the paper will show whatever is in the PAPERTEXT05 definition.
- Added hint text to tell the player to hold 'use' to keep viewing the
paper
- Added the ability to also choose different papers and fonts, via the
2nd and 3rd arguments of the script.
- Added a selection of bitmap fonts to test functionality, and some
quick alternate paper backgrounds (they're not as battered, so feel free
to replace - or use them in offices or something).

I've set up a 'demo mode' for now on the starting paper in C3M0_A...  It
will pick a new font and a new paper type every time you hit 'use' on
it, so you can see the full effect.  For other papers, just set the
papertype properly, and the 'demo mode' will go away - I'll remove it in
a few days altogether.