Allow matching TextPaper sprites to paper type
- ZScriptified the Textpaper actor and added some handling so that it
changes its sprite based on the papertype that is set in the script
arguments.  Sprites must be in the same order as the larger paper
backgrounds.
- Note that this will require a new sprite be created for each new
papertype that might be added...  I also renamed the paper base images
so that we can have more than 9 paper backgrounds (keeping the name
under 8 characters).

I'm assuming that we'll want some graphical "paper" one-offs
occasionally as well - this same code could be used to 'look at'
pictures, items, etc... Just leave the message number as zero and the
"paper" image will show up by itself.