Material tests (check TEST_ACT for a demo)
I was playing around with the feature now and it took me around 2 hours
to get used to it and create 4 different types of textures.

I think the brick, vent and rock variants work well, but I also noticed
that textures with details like doors or computers need definitely
custom work to look good. It's nothing that can be automated when it
comes to "complex" textures, but it can most propably be automated a lot
when it comes to normal patterns.

I am not sure if we can manage this alone or if we need someone who is
skilled with stuff like that and who is able to do this for the textures
which benefit from the new feature.

Hard to decide, looking forward to your thoughts on this