C3M0_A: Kill reward script fixes/updates
- Reworked the 403 script as a KILL script so we don't have to override
every thing's action special.
- Added "Zombie" species to all Zombies, and added check for that
Species (*or* the user_lightburns UDMF property) in the KILL script so
that only those actors count as 'undead'.
- Added check for the player as the killer so you only get rewarded if
you actually killed the zombie yourself (and added check for Lightning
so the player gets credit for the kill with Tesla gun)
- Simplified the if statments in the reward script to a switch statement
- Tweaked count hadnling and display so that this *should* actually be
multiplayer friendly (not that it matter here, but...)
- Added logic to use 'night' vs. 'nights' if nightcounter is less than 1
- Moved nightcounter++ call to a different place in the code so that it
only increments at the very end of each day cycle.