C3M1: ACS and crosshair tweaks
- Changed user_corsshair handling so that the custom crosshair only
shows up if the line they are set on is a script trigger or has a UDMF
lock number assigned (the assumption being that we won't need to prompt
the player to activate any other type of line).  I can add another
property to override this behavior if needed.
- Re-worked the rusty lock script to take advantage of this behavior and
simplify the logic - basically, I use the script to handle the shovel
inventory check and lock-breaking.  The end of that script then directly
sets the line special to poly_doorswing - since this isn't a
script-running special, the user_crosshair image goes away.
- Also re-worked the C4 pickup script - the C4 activation line starts
out with no special, so no user_crosshair shown, then the C4 pickup
script sets the line special to the correct values - a script call - so
the user_crosshair shows up.