C3M1, LANGUAGE, ZScript: Various small fixes/additions
- Fixed inside door of first prisoner barracks (seen in intro cutscene)
- Restored floor reflectivity in the prisoners' underground tunnel so it
actually looks wet
- Updated line 19986 properties and tweaked script for the 'opening'
gate in between the western and easter camp portions (silences map load
warning)
- Fixed double-sided line 1895 with no back side (silences map load
warning)
- Added ZScript handling so that player weapon can be fired via ACS;
updated shovel gate script so that if you 'use' the locked gate while
you have the shovel, you'll raise the shovel (if you didn't have it as
current weapon) and hit the gate with it instead of the line just
changing.
- Tweaked ZScript Base.InPlayerSight functoon to be static, and added a
script with checks for not being in sight of player to wrap the
Thing_Remove call at the end of the prisoner line so that prisoners
wouldn't just disappear into thin air in front of you.
- Fixed LANGUAGE strings that referred to the eastern side of camp when
they meant the western side
- Slight optimization of hold-to-use code, and fix to frame setting of
TextPaper actor
- Tweaked the blur shader to be less pixelated at higher blur values