Added dummy actor/models for driveable Sherman tank
MODELS, MATERIALS, DECORATE, DEFS & MAPS
:new: Added new US_Sherman actor, which is planned to be a driveable
vehicle!
:new: The new actor is provided even with fitting materials (model from
CoD2), except that sherman_d.png skin needs to be fixed a bit (later)
:new: Added above actor inside TEST_ACT map, which now I have moved the
PlayerStart inside Lights room

:question: @AFADoomer Is it possible for you to come with a fancy actor
for a driveable tank, similar to something like this (
https://www.youtube.com/watch?v=DabpM9kxYL0 ) ? I think we'll need to
morph the Player for such effect, since we shouldn't be affected by
tilt, flinch and sprint properties. You may notice that I have split the
turret from the model... Hope it isn't a big request, but we would be
very excited to have something like this and include on a city-themed
map for C3.