Driveable Tank Update
- Tweaked offsets on turret and cannon to set the center of rotation to
the correct location
- Moved canon spawn handling to turret's PostBeginPlay function and set
up the gun to warp its location to the correct offset every tic
- Updated removal code so that the cannon disappears when you +use the
US_Sherman actor, and when you un-morph
- Simplified some of the movement code that was unnecessarily
complicated
- Added 'alternate' driving style...  Hold down +run (Shift) while
driving, and the tracks will auto-steer until they align with the turret
angle.
- Added internal handling so that turret and gun move properly in
conjunction with the treads when the treads are pitched/rolled
- Tweaked status bar so that the aiming reticle doesn't draw over the
automap

TODO:
- Add handling to cause pitch/roll changes as terrain changes
- Tweak movement and turn speeds (movement is too fast right now)
- Make min/max pitch handling less glitchy...  The pitch can get 'stuck'
if you go to min/max and keep going.
- Finalization of Alt Fire
- Tweak camera position by offsetting slightly from the maximum
distance, so that you don't get flickering walls in your vision if the
camera is direclty on a wall line.
- Sounds