C3M1: Final fixes, should be playable now
@Ozymandias81

- I noticed that the "ZyklonBSteamSpawners" aren't working anymore, can
you check whats wrong?
- The cyn door still doesn't use the "cyan texture" even though it's
correctly applied in the editor. Also I can't open it, doesn't work
(reuse is also impossible). Do you have an idea Mr. Bugfixer? :)
- Noticed that the death sequence of the Schabbs Jr. misses a frame, I
guess it's physically there in the folder but it is not in the code (the
Doc walls too quick when killed)
- Can you also make the metal debris larger? They are simply not visible
after the detonation of the refinery when dropping down. At least twice
as large (the dropping ones)