C3M3: Uploaded some little fixes and encountered an issue with the skybox
@AFADoomer - When you reach the roof of the flak tower, the skybox was
working well in the past but it doesn't work well anymore now. It seems
that the static skybox thing you've created doesn't take z-transitions
through portals anymore. Maybe it has something to do with the "anchor"
changes you did for C3M5_C (Airship map). Do you have an idea how this
can be fixed? Is it a bug or is there a workaround to make it work
again?

Thanks already for taking a look into this.