Driveable Tank Updates
- Split the player tank classes out into their own file (tank.txt)
- Tweaked base class handling of tank fear to only affect actors that
move around (Speed > 0).
- Added handling to offset the tank's speed so that cl_run has no effect
and holding 'run' (shift) doesn't make you go faster
- Reversed logic for changing into auto-steer mode and added an
"autosteer" cvar that swaps the behavior so that we can make this user
configurable if we want (and because I like it the other way :tongue: ).
- Set ShermanPlayer's Rocket DamageFactor to 0.5 so rockets hurt less

TODO: Need to slow down the cannon firing rate, which is easy, but I
also want to add handling to allow firing both primary and alternate at
the same time (code is mostly complete, just can't finish it tonight).