Initial work on ZScriptified Tank enemies
- Added Tank class that is slope-aware and supports multi-part tank
actors (treads, turret, gun)
- Added newSSTank1 actor that emulates SSTank1 behavior, but with a
moving turret (uses the Sherman tank model pieces for now)
- Fixed slope handling to keep the tank on the ground when the actual
actor is hovering over the edge of a slope

Please try summoning the newSSTank1 actor and let me know if there are
significant issues with it as an enemy.  I already know that the Sherman
pieces don't get removed when it dies, yet - still todo on that.