C3M3, Skybox Viewpoint Fix
- Fixed some logic in the height check and offset calculation...  Not
sure what I was thinking before.
- Added a proper anchorpoint to better align the skybox
- Tweaked the flaktower roof sectors' heights - since the
skyboxviewpoint z-height is calculated based on the player's z position
in relation to the anchor point, the sectors needed to be raised by ~700
units or so to raise the view above the 'ground' of the skybox sector.
With normal geometry, this happens as a matter of course, but since the
roof is accessed via a portal, the offset wasn't what we would expect
from being on top of a tower.

It may be better to move the flaktower roof portion "west" on the map,
so that it is centered on the town (and the anchorpoint - TID 100), so
that the burning city is in the right place in relation to the tower.

I may be able to further tweak things so that we can us ACS to script
changing the skybox anchorpoint...  But that'll take a little bit of
work, I think.