Turret Fixes
- Re-worked player positioning/aiming to use a different implementation
- Fixed bullets hitting player when positioned at ~45 degree angle from
the gun
- Fixed incorrect light color variable

The turret in C3M2 may need to have its position adjusted, or a section
of the wall in front of it lowered.  Right now, it can't aim below
horizontal because the barrel is basically level with the top of the
wall.  The old KTurrets had +NOGRAVITY, so could float in mid-air to add
height - these don't have that (because it was really obvious if the
feet weren't on the ground).

The player can also end up 'hopping' in place because you end up
standing on the edge of the slopes as you are warped to maintain
position behind the gun...  I'll have to think about how to fix that.