Add a version of the old video shader that works with textures
Add OldVideoTx.fp, which is a version of the Old Video shader (OldVideo.fp) that can be applied to textures.
Add examples to boashaders.txt and TEST_DLG.

Also, if someone can figure out how to add the shader effects to the alpha channel of a texture, that would be even more awesome, as it would allow mappers to place "textures/briefing/oldvideo.png" in front of any cinema/projector screens.