Tank updates
- Added scaled-back Trample effect to all tanks.  Configurable with the
VehicleBase.TrampleRadius property (defaults to the actor radius in
size, so probably OK as-is for all existing actors)
- Also added 'FlattenableProp' base class and set the rye grass actor to
inherit from it.  This causes the rye grass to be flattened to the
ground when a player or any actor inheriting from the Base class runs
into/over it or it is damaged by an explosion (with appropriate
directional handling for explosions and movement).  Just run through the
rye field at the beginning of C3M2 to see it in action (or watch the
tank drive in the grass).