Tank POV camera experiment
- Implemented a "gunner's point of view" camera for the Sherman...
Toggle the chase camera ('chase' in console) while you are driving a
tank to see it.
- Also fixed using shift to swap back to the primary driving mode when
the autosteer cvar is set

This doesn't really give any advantage other than a faster turret turn
speed, and the effect's not perfect.  Not sure I like it.  But, it's
there...  I can clean up the activation so it's not dependent on the
chasecam cheat if we want to use it at all (or leave it as-is as an
easter egg of sorts).