Multi-part Tank Actors
- Re-implemented SSTank1-SSTank4 and T34Tank as multi-part actors with
moving turrets/guns.  The AI is similar to the old logic, but not quite
the same...  I anticipate that there will need to be some tweaking to
their behavior.
- In order to facilitate doing this, I made several configurable base
actor classes so that adding new tanks/vehicles will be easier in the
future...
- Also fixed the Sherman player class so that the movement speed isn't
affected by cl_run or holding shift.
- Also did a minor bug fix for the base debris actor

Still to do:  SSTank3 needs its cannon split out as a separate model.
SSTank4's turret-mounted gun could be split out too, but that's probably
overkill.  All of the tanks (except the Sherman) need at least one
texture variant with their wheels/treads in an alternate position so we
can animate the skin when the tank is moving (everything's in place
except the alternate frames' Skin assignments in MODELDEF, which are
commented out).  There's still a little bit of cleanup to do, but I need
to get this committed...