C3M2: Added more outside areas of the second battlefield, bringing it closer to a finished state
@AFADoomer - Considering that the Wehrmacht Panzerschreck enemies are a
real treat to your Sherman, do you have an idea how to make these
enemies more exposed to the player, more "seeable", not an exclamation
mark but something else that makes it clear to to that you need to kill
them instantly as your tank is in danger? (only make that visible when
you are "in tank")

Next to that, some kind of "Tank Repair Inventory Item" might be useful,
just as the field kit, just for the tank, and for 20% of tank health.

@Ozymandias81 - Currently the Panzerschreck soldiers are aiming perfect
(shoot = hit). Can you give them a slight chance to miss their target,
shooting next to them instead hitting them? (e.g. 50% chance of good
aim, 50% of bad aim)