Tank tweaks
- Added 'reverse' driving for tank enemies so they back up and turn
around if they hit a wall instead of just grinding themselves into it,
and don't spin in place if they're supposed to be a wheeled vehicle
(e.g, SSTank2 and SSTank4)
- Minor tweaks to gun pitch limits of SSTank3 and radius of
FriendlySherman, and fixed tank line-of-sight checks to work more
correctly
- Added handling so that tanks will still try to face their target while
the turret is waiting to fire...  Basically, this lets the tanks
actually try to use their forward machine gun more often so it's not a
wasted effect.
- Fixed shift still having an effect on tank player's run speed in some
situations.
- Added drawing of inventory bar to the tank HUD...  I don't know if we
can disable use of inventory while  in the tank.
- Also did a minor tweak to the minesweeper on-screen drawing scaling