Various edits to ENU, and new pickup message for AktenDreamClue
Re-write the last two stages of C3M1 briefing to be shorter, and include info about Lt. Arlow's special job for you.
Begin work on C3M2 briefing.
Add a special pickup message for the AktenDreamClue actor because it's weird seeing "..?" outside of the dream.
Also, fix location error in EPILOGUEC2 as noted by wolfmanfp on the ZDoom forums thread.

@Tormentor667: I've tentatively named the "zombie" operation Speerkrieg, but I don't know German at all. Feel free to change the name. Also, I'm thinking of renaming C3M2 to "Operation: Inferno", and having Sgt. Nash stand beside the truck instead of Lt. Ascher.