C3M5_C: Added some rooms at the beginning and also did some plans regarding the plot of this map.
In C3M5_B, BJ should get into the air ship in disguise together with
Ascher. First I want some puzzle on the airship, then destroying the
engine, then action, then crash and then we proceed to C3M6_A.

For this I need "Ascher" in Nazi Officer uniform (incl. mugshot,
@Ozymandias81 ) as well as a crash/tilt sequence that we can implement
somehow which seems to be a bit complicated now with the tilt stuff so I
hope @AFADoomer you can help here.

After the engine is destroyed, the tilting should start and the airshop
should slowly go down into the mountains and at some point it crashes
and the map ends (my plan is that the player has to enter a certain area
of the ship before the crash (countdown) where he is safe and will
survive the crash). Maybe you have an idea how to script that properly
(countdown, tilt, crash)