Heat Shader
- Added heat shimmer shader.
- Can be given manually with the HeatShaderControl inventory item
(similar to blur - add 1 item per tick desired duration), or use the
HeatEffectGiver actor...
- The HeatEffectGiver actor's sprite is a flatsprite circle in GZDB;
scale the circle (set x and y scale in editor) to match the size of the
sprite to the radius at which you want the player to start seeing heat
shimmer.  Shimmer intensifies as the player gets closer to the center of
the HeatEffectGiver actor.

This is not really realistic because it's a full-screen effect only...
A better effect would be if we could only shimmer the part of the screen
with the heat source below it.