Heat Shader Update
- Updated/redid shader code to change scale of effect and add blur
- Also (mostly) fixed some artifacting that occurred as the effect ran
for a longer time
- Tweaked the HeatEffectGiver to automatically fill the vertical space
above/below where it is placed.

NOTE: When using the HeatEffectGiver, they use the actual circular
radius around the point as the boundary of the heat effect (not the
usual square boundary), so you can end up with some 'holes' at the edges
of heat zones if the area isn't circular and you don't make the heat
giver large enough.