Heat Shader Tweaks
- Re-worked the heat shader again to slow things down and increase the
'scale' of the effect.  I think this is closer to what you want, Torm?
- Also added some randomization so the patterns are less obvious, and
fixed some issues with the left side of the screen not being distorted
- Also added some comments to the various sections of code
- Tweaked the HeatEffectGiver so it should be able to be
Activated/Deactivated, and to scale the amount of the shader effect
given properly between 0 and 128, regardless of the size of the giver
object.

This last change allows for actors (such as FireSpawner) to just inherit
directly from the HeatEffectGiver, simplifying the DECORATE hackery
required to make them give the heat shader control object.  Hopefully
this fixes the crashes that were occuring with the previous method.